requirements.txt
fastapi
uvicorn
docling
python-multipart
pymupdf
pytesseract
pillow
python-magic

